Move handleCreateIntentCallback to view model
